+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- abi `codegen` support for `cosmos/Ethermint` (#1936)
+### Fixed
+- Ensure non evm custom datasource do not generate abis (#1936)
 
 ## [3.4.0] - 2023-08-10
 ### Added